+++ src/conditions/base/evm.ts
@@ -10,7 +10,7 @@ export const STANDARD_CONTRACT_TYPES = ['ERC20', 'ERC721'];
 export class EvmCondition extends Condition {
   public readonly schema = Joi.object({
     contractAddress: Joi.string().pattern(ETH_ADDRESS_REGEXP).required(),
-    chain: Joi.string()
+    chain: Joi.number()
       .valid(...SUPPORTED_CHAINS)
       .required(),
     standardContractType: Joi.string()