+++ src/conditions/base/index.ts
@@ -1,4 +1,4 @@
 export { Condition } from './condition';
 export { EvmCondition } from './evm';
 export { RpcCondition } from './rpc';
-export { TimelockCondition } from './timelock';
+export { TimeCondition } from './time';